Moved a bunch of shared definitions into their own file
